refactor: move news request logic from list to page; add aler if status === rejected
