+++ src/components/News/NewsList.jsx
@@ -1,34 +1,34 @@
-import { useEffect } from "react";
+// import { useEffect } from "react";
 
-import { useSelector, useDispatch } from "react-redux";
-import { getNewsThunk } from "../../redux/news/operationsNews";
-import { selectNews } from "../../redux/news/selectorNews";
+// import { useSelector, useDispatch } from "react-redux";
+// import { getNewsThunk } from "../../redux/news/operationsNews";
+// import { selectNews } from "../../redux/news/selectorNews";
 
-import Alert from "../Alert.jsx";
-import Loader from "../Loader";
+// import Alert from "../Alert.jsx";
+// import Loader from "../Loader";
 import NewsItem from "./NewsItem";
 
-const NewsList = () => {
-  const dispatch = useDispatch();
-  const { news, status, error } = useSelector(selectNews);
-  const alert = { text: "", type: "" };
+const NewsList = ({ news }) => {
+  // const dispatch = useDispatch();
+  // const { news, status, error } = useSelector(selectNews);
+  // const alert = { text: "", type: "" };
 
-  useEffect(() => {
-    dispatch(getNewsThunk());
-  }, [dispatch]);
+  // useEffect(() => {
+  //   dispatch(getNewsThunk());
+  // }, [dispatch]);
 
-  if (status === "loading") {
-    return <Loader />;
-  }
+  // if (status === "loading") {
+  //   return <Loader />;
+  // }
 
-  if (status === "rejected") {
-    alert.text = `${error}. Please contact your administrator!`;
-    alert.type = "danger";
-  }
+  // if (status === "rejected") {
+  //   alert.text = `${error}. Please contact your administrator!`;
+  //   alert.type = "danger";
+  // }
 
   return (
     <>
-      {error && <Alert state={alert} />}
+      {/* {error && <Alert state={alert} />} */}
       <ul>
         {news.map((newsItem) => (
           <li className="mb-3" key={newsItem.id}>
