+++ src/pages/NewsPage.jsx
@@ -1,20 +1,51 @@
+import { useEffect } from "react";
 import { Link } from "react-router-dom";
 
+import { useSelector, useDispatch } from "react-redux";
+import { getNewsThunk } from "../redux/news/operationsNews";
+import { selectNews } from "../redux/news/selectorNews";
+
+import Alert from "../components/Alert";
+import { useAlert } from "../assets/utils/useAlert";
+
+import Loader from "../components/Loader";
 import NewsList from "../components/News/NewsList";
 import setPageTitle from "../assets/utils/setPageTitle";
 
 const NewsPage = () => {
   setPageTitle("News");
+  const dispatch = useDispatch();
+  const { news, status, error } = useSelector(selectNews);
+  const { alert, showAlert } = useAlert();
+
+  useEffect(() => {
+    dispatch(getNewsThunk());
+  }, [dispatch]);
+
+  useEffect(() => {
+    if (status === "rejected") {
+      showAlert(`${error}. Please contact your administrator!`, "danger");
+      return;
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [status]);
+
+  if (status === "loading") {
+    return <Loader />;
+  }
 
   return (
-    <section className="py-4">
-      <NewsList />
-      <div className="mt-3">
-        <Link className="btn btn-primary" to={"/news/new"}>
-          new news
-        </Link>
-      </div>
-    </section>
+    <>
+      {error && <Alert state={alert} />}
+      <section className="py-4">
+        <NewsList news={news} />
+        <div className="mt-3">
+          <Link className="btn btn-primary" to={"/news/new"}>
+            new news
+          </Link>
+        </div>
+      </section>
+    </>
   );
 };
